[DevTSAN] Treat each work item as a thread for GPU device
* Update tests to make it more robust
