+++ sycl/test-e2e/ThreadSanitizer/check_shared_usm.cpp
@@ -1,0 +1,22 @@
+// REQUIRES: linux, cpu || (gpu && level_zero)
+// ALLOW_RETRIES: 10
+// RUN: %{build} %device_tsan_flags -O0 -g -o %t.out
+// RUN: %{run} %t.out 2>&1 | FileCheck %s
+#include "sycl/detail/core.hpp"
+#include "sycl/usm.hpp"
+
+int main() {
+  sycl::queue Q;
+  auto *array = sycl::malloc_shared<char>(1, Q);
+
+  Q.submit([&](sycl::handler &h) {
+     h.parallel_for<class Test>(sycl::nd_range<1>(128, 8),
+                                [=](sycl::nd_item<1>) { array[0]++; });
+   }).wait();
+  // CHECK: WARNING: DeviceSanitizer: data race
+  // CHECK-NEXT: When write of size 1 at 0x{{.*}} in kernel <{{.*}}Test>
+  // CHECK-NEXT: #0 {{.*}}check_shared_usm.cpp:[[@LINE-4]]
+
+  sycl::free(array, Q);
+  return 0;
+}
