+++ sycl/test-e2e/ThreadSanitizer/check_unaligned_access.cpp
@@ -15,7 +15,7 @@ int main() {
   auto *array = sycl::malloc_device<S>(1, Q);
 
   Q.submit([&](sycl::handler &h) {
-     h.parallel_for<class Test>(sycl::nd_range<1>(32, 8),
+     h.parallel_for<class Test>(sycl::nd_range<1>(128, 8),
                                 [=](sycl::nd_item<1>) { array[0].x++; });
    }).wait();
   // CHECK: WARNING: DeviceSanitizer: data race
